add urdf option
