--- nautilus_worlds/worlds/underwater.world
@@ -1,113 +1,0 @@
-
-<?xml version="1.0" ?>
-<sdf version="1.5">
-  <world name="default">
-
-    <include>
-      <uri>model://ground_plane</uri>
-    </include>
-
-    <include>
-      <uri>model://sun</uri>
-    </include>
-
-    <model name="ceiling_plane">
-      <static>true</static>
-      <pose>0 0 40 0 0 0</pose>
-      <link name="link">
-        <collision name="collision">
-          <geometry>
-            <box>
-              <size>200 200 .1</size>
-            </box>
-          </geometry>
-        </collision>
-        <visual name="visual">
-          <cast_shadows>false</cast_shadows>
-          <geometry>
-            <box>
-              <size>200 200 .1</size>
-            </box>
-          </geometry>
-          <material>
-            <script>
-              <uri>file://media/materials/scripts/gazebo.material</uri>
-              <name>Gazebo/BlueTransparent</name>
-            </script>
-          </material>
-        </visual>
-
-        <visual name="visual_sideA">
-          <pose>100.05 0 -20 0 0 0</pose>
-          <cast_shadows>false</cast_shadows>
-          <geometry>
-            <box>
-              <size>0.1 200 39.9</size>
-            </box>
-          </geometry>
-          <material>
-            <script>
-              <uri>file://media/materials/scripts/gazebo.material</uri>
-              <name>Gazebo/BlueTransparent</name>
-            </script>
-          </material>
-        </visual>
-
-        <visual name="visual_sideB">
-          <pose>-100.05 0 -20 0 0 0</pose>
-          <cast_shadows>false</cast_shadows>
-          <geometry>
-            <box>
-              <size>0.1 200 39.9</size>
-            </box>
-          </geometry>
-          <material>
-            <script>
-              <uri>file://media/materials/scripts/gazebo.material</uri>
-              <name>Gazebo/BlueTransparent</name>
-            </script>
-          </material>
-        </visual>
-
-        <visual name="visual_sideC">
-          <pose>0 100.05 -20 0 0 0</pose>
-          <cast_shadows>false</cast_shadows>
-          <geometry>
-            <box>
-              <size>200 0.1 39.9</size>
-            </box>
-          </geometry>
-          <material>
-            <script>
-              <uri>file://media/materials/scripts/gazebo.material</uri>
-              <name>Gazebo/BlueTransparent</name>
-            </script>
-          </material>
-        </visual>
-
-        <visual name="visual_sideD">
-          <pose>0 -100.05 -20 0 0 0</pose>
-          <cast_shadows>false</cast_shadows>
-          <geometry>
-            <box>
-              <size>200 0.1 39.9</size>
-            </box>
-          </geometry>
-          <material>
-            <script>
-              <uri>file://media/materials/scripts/gazebo.material</uri>
-              <name>Gazebo/BlueTransparent</name>
-            </script>
-          </material>
-        </visual>
-
-      </link>
-    </model>
-
-    <include>
-      <uri>model://nautilus</uri>
-      <pose>0 0 1 0 0 0</pose>
-    </include>
-
-  </world>
-</sdf>
