+++ src/nautilus_description/urdf/model.gazebo
@@ -1,0 +1,19 @@
+<?xml version="1.0"?>
+<robot>
+  <!-- <gazebo reference="body"> -->
+    <!-- <plugin name="BuoyancyPlugin" filename="libBuoyancyPlugin.so"> -->
+        <!-- <fluid_density>1000</fluid_density> -->
+        <!-- <link name="body"> -->
+        <!-- TODO: calculate volume of ROV -->
+        <!-- TODO: provide center of mass -->
+          <!-- <volume>0.0215755</volume> -->
+          <!-- <volume>0.05</volume> -->
+          <!-- <center_of_volume>0 -.135 0</center_of_volume> -->
+        <!-- </link> -->
+      <!-- </plugin> -->
+  <!-- </gazebo> -->
+  
+  <!-- <gazebo>
+    <plugin name="thruster" filename="libthruster.so"/>
+  </gazebo> -->
+</robot>
