+++ src/nautilus_worlds/package.xml
@@ -1,0 +1,35 @@
+<?xml version="1.0"?>
+<package>
+  <name>nautilus_worlds</name>
+  <version>0.0.0</version>
+  <description>The nautilus_worlds package. Powers the simulation of nautilus via gazebo,
+               Credit to PalouseRobosub for inspiration (https://github.com/PalouseRobosub/robosub_simulator)</description>
+
+  <maintainer email="gunarp@uw.edu">Peter Gunarso</maintainer>
+
+  <license>TODO</license>
+
+  <build_depend>gazebo_ros</build_depend>
+  <run_depend>gazebo_ros</run_depend>
+  <build_depend>roscpp</build_depend>
+  <run_depend>roscpp</run_depend>
+  <build_depend>rospy</build_depend>
+  <run_depend>rospy</run_depend>
+  <build_depend>gazebo_ros</build_depend>
+  <run_depend>gazebo_ros</run_depend>
+  <build_depend>gazebo_plugins</build_depend>
+  <run_depend>gazebo_plugins</run_depend>
+  <build_depend>gazebo_msgs</build_depend>
+  <run_depend>gazebo_msgs</run_depend>
+  <build_depend>std_msgs</build_depend>
+  <run_depend>std_msgs</run_depend>
+  <build_depend>message_generation</build_depend>
+  <run_depend>message_generation</run_depend>
+  <buildtool_depend>catkin</buildtool_depend>
+
+  <!-- The export tag contains other, unspecified, tags -->
+  <export>
+    <!-- Other tools can request additional information be placed here -->
+    <gazebo_ros gazebo_plugin_path="${prefix}/../../devel/lib" />
+  </export>
+</package>
